chore: #192 read api test k6 스크립트 정리
