+++ k6/test/api-load-test/read/1-check-group-participating-v1-k6-test.js
@@ -1,0 +1,38 @@
+import { sleep } from 'k6';
+import {check} from 'k6';
+import { SharedArray } from 'k6/data';
+import {checkGroupParticipating} from "../../../common/api/api-call/api-call.js";
+import {parseResponseBody} from "../../../common/api/util/api-util.js";
+
+const tokens = new SharedArray('tokens', () => JSON.parse(open('../../../secret/tokens.json')));
+
+export const options = {
+    setupTimeout: '30m',
+    scenarios: {
+        default: {
+            executor: 'per-vu-iterations',
+            vus: 100,
+            iterations: 1,
+            maxDuration: '30m',
+        },
+    },
+};
+
+export function setup() {
+    console.log("⏰ 5초 대기 시작.");
+    sleep(5);
+    console.log("✅ 5초 대기 완료.\n");
+}
+
+export default function () {
+    const vuIndex = __VU - 1;
+    const token = tokens[vuIndex];
+
+    const res = checkGroupParticipating(token);
+    const responseData = parseResponseBody(res).data;
+
+    check(res, {
+        'API 응답 상태 코드 200': (r) => r.status === 200,
+        '응답 데이터 - checkParticipating 존재': () => responseData?.checkParticipating !== undefined
+    });
+}
