--- k6/test/api-load-test/read/1-check-group-participating-v1/script/set-up.js
@@ -1,10 +1,0 @@
-import {insertData} from "../../../../../common/db/util/data-insert-runner.js";
-import {
-    createDummyData,
-} from "../../../../../common/db/data/set-up-data/set-up-data-1.js"
-
-async function setUp() {
-    await insertData(createDummyData);
-}
-
-setUp();
