--- k6/test/api-load-test/read/1-check-group-participating-v1/script/tear-down.js
@@ -1,7 +1,0 @@
-import {removeAllTestDataCreateToday} from "../../../../../common/db/util/data-clear-runner.js";
-
-async function tearDown() {
-    await removeAllTestDataCreateToday();
-}
-
-tearDown();
