+++ k6/test/api-load-test/read/3-get-my-todos-v1-k6-test.js
@@ -1,12 +1,13 @@
 import { sleep } from 'k6';
 import {check} from 'k6';
 import { SharedArray } from 'k6/data';
-import {getMyDailyTodos} from "../../../../../common/api/api-call/api-call.js";
-import {parseResponseBody} from "../../../../../common/api/util/api-util.js";
+import {getMyDailyTodos} from "../../../common/api/api-call/api-call.js";
+import {parseResponseBody} from "../../../common/api/util/api-util.js";
 
-const tokens = new SharedArray('tokens', () => JSON.parse(open('../../../../../secret/tokens.json')));
-const data = new SharedArray('temp', () => [JSON.parse(open('./data.json'))])[0];
-const challengeGroupIds = data.groupIds;
+import {getChallengeGroupIdsPerMember} from "../../../common/db/data/set-up-data/read-test/read-test-set-up-data-1.js";
+// import {getChallengeGroupIdsPerMember} from "../../../common/db/data/set-up-data/read-test/read-test-set-up-data-2.js";
+
+const tokens = new SharedArray('tokens', () => JSON.parse(open('../../../secret/tokens.json')));
 
 export const options = {
     setupTimeout: '30m',
@@ -21,16 +22,21 @@ export const options = {
 };
 
 export function setup() {
-    console.log("⏰ 10초 대기 시작.");
-    sleep(10);
-    console.log("✅ 10초 대기 완료.\n");
+    const challengeGroupIds = getChallengeGroupIdsPerMember();
+    const todayDateInKst = getCurrentDateInKst();
+
+    console.log("⏰ 5초 대기 시작.");
+    sleep(5);
+    console.log("✅ 5초 대기 완료.\n");
+
+    return {challengeGroupIds, todayDateInKst};
 }
 
-export default function () {
+export default function (data) {
     const vuIndex = __VU - 1;
     const token = tokens[vuIndex];
-    const challengeGroupId = challengeGroupIds[vuIndex][0];
-    const todayDate = getCurrentDateInKst();
+    const challengeGroupId = data.challengeGroupIds[vuIndex][0];
+    const todayDate = data.todayDateInKst;
 
     const res = getMyDailyTodos(token, challengeGroupId, todayDate);
     const responseData = parseResponseBody(res).data;
@@ -39,11 +45,6 @@ export default function () {
     check(res, {
         'API 응답 상태 코드 200': (r) => r.status === 200
     });
-}
-
-export function teardown() {
-    console.log("🧹 5초 후 테스트 데이터 정리 시작.");
-    sleep(5);
 }
 
 function getCurrentDateInKst() {
