--- k6/test/api-load-test/read/3-get-my-todos-v1/run.sh
@@ -1,3 +1,0 @@
-node ./script/set-up.js
-k6 run ./script/k6-test.js
-node ./script/tear-down.js
