--- k6/test/api-load-test/read/3-get-my-todos-v1/script/tear-down.js
@@ -1,7 +1,0 @@
-import {removeAllTestDataCreateToday} from "../../../../../common/db/util/data-clear-runner.js";
-
-async function tearDown() {
-    await removeAllTestDataCreateToday();
-}
-
-tearDown();
