--- k6/test/api-load-test/read/4-get-todo-certifications-for-reviewer-v1/run.sh
@@ -1,3 +1,0 @@
-node ./script/set-up.js
-k6 run ./script/k6-test.js
-node ./script/tear-down.js
