--- k6/test/api-load-test/read/4-get-todo-certifications-for-reviewer-v1/script/tear-down.js
@@ -1,43 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-import {
-    deleteAllMemberById,
-    deleteDailyTodoStatsByMemberId,
-    deleteNotificationTokenByMemberId
-} from "../../../../../common/db/query/member-query.js";
-import {createLocalDbConnection, createSshTunnelDbConnection} from "../../../../../common/db/util/db-util.js";
-
-const temp = JSON.parse(fs.readFileSync(path.join('./script/temp.json'), 'utf-8'));
-const connection = await createLocalDbConnection(); // Local DB 커넥션
-// const connection = await createSshTunnelDbConnection(); // AWS DB 커넥션
-
-async function tearDown() {
-    const {
-        firstTestMemberId,
-        testMemberDataSize
-    } = temp;
-
-    const targetMemberIds = [];
-    for (let i = 0; i < testMemberDataSize; i++) {
-        targetMemberIds.push(firstTestMemberId + i);
-    }
-
-    try {
-        await connection.beginTransaction();
-
-        await deleteDailyTodoStatsByMemberId(connection, targetMemberIds);
-        await deleteNotificationTokenByMemberId(connection, targetMemberIds);
-        await deleteAllMemberById(connection, targetMemberIds);
-
-        await connection.commit();
-        console.log('\n🥳 테스트 데이터 삭제 완료!\n');
-    } catch (error) {
-        await connection.rollback();
-        console.error(`❌ 에러 발생! 롤백 수행됨.`);
-        console.error(error);
-    } finally {
-        await connection.end();
-    }
-}
-
-tearDown();
