--- k6/test/api-load-test/read/6-get-todo-histories-v1/script/tear-down.js
@@ -1,7 +1,0 @@
-import {removeAllTestDataCreateToday} from "../../../../../common/db/util/data-clear-runner.js";
-
-async function tearDown() {
-    await removeAllTestDataCreateToday();
-}
-
-tearDown();
