--- k6/test/api-load-test/read/7-get-my-group-activity-info-v1/script/tear-down.js
@@ -1,7 +1,0 @@
-import {removeAllTestDataCreateToday} from "../../../../../common/db/util/data-clear-runner.js";
-
-async function tearDown() {
-    await removeAllTestDataCreateToday();
-}
-
-tearDown();
