+++ k6/test/api-load-test/read/9-get-my-profile-v1-k6-test.js
@@ -1,0 +1,38 @@
+import { sleep } from 'k6';
+import {check} from 'k6';
+import { SharedArray } from 'k6/data';
+import {getMemberProfile} from "../../../common/api/api-call/api-call.js";
+import {parseResponseBody} from "../../../common/api/util/api-util.js";
+
+const tokens = new SharedArray('tokens', () => JSON.parse(open('../../../secret/tokens.json')));
+
+export function setup() {
+    console.log("⏰ 5초 대기 시작.");
+    sleep(5);
+    console.log("✅ 5초 대기 완료.\n");
+}
+
+export const options = {
+    setupTimeout: '30m',
+    scenarios: {
+        default: {
+            executor: 'per-vu-iterations',
+            vus: 100,
+            iterations: 1,
+            maxDuration: '30m',
+        },
+    },
+};
+
+export default function () {
+    const vuIndex = __VU - 1;
+    const token = tokens[vuIndex];
+
+    const res = getMemberProfile(token);
+    const responseData = parseResponseBody(res).data;
+
+    // TODO : 검증 로직 추가
+    check(res, {
+        'API 응답 상태 코드 200': (r) => r.status === 200
+    });
+}
